test: move many-to-many junction cascade test to functional suite
Relocate the #11738 regression test out of test/github-issues into the
functional many-to-many suite and run it against all configured drivers
instead of better-sqlite3 only. Assertions account for Spanner (no
UPDATE/DELETE CASCADE) and Oracle (no UPDATE CASCADE), matching the
junction builder's per-driver fallbacks.
